adapted to moved SimuComConfig
