+++ bundles/de.uka.ipd.sdq.dsexplore/src/de/uka/ipd/sdq/dsexplore/helper/ConfigurationHelper.java
@@ -11,7 +11,7 @@ import org.eclipse.debug.core.ILaunchConfigurationWorkingCopy;
 import org.palladiosimulator.analyzer.workflow.core.ConstantsContainer;
 import org.palladiosimulator.solver.core.models.PCMInstance;
 
-import de.uka.ipd.sdq.simucomframework.SimuComConfig;
+import de.uka.ipd.sdq.simucomframework.core.SimuComConfig;
 
 /**
  * helps handling the {@link ILaunchConfiguration} needed during a Design Exploration. Offers methods to generate new {@link ILaunchConfiguration} to run the simulation and also delete the generated {@link ILaunchConfiguration}s again.
