+++ bundles/de.uka.ipd.sdq.pcm.cost/src-man/de/uka/ipd/sdq/pcm/cost/helper/CostUtil.java
@@ -10,7 +10,7 @@ import de.uka.ipd.sdq.pcm.cost.VariableProcessingResourceCost;
 import de.uka.ipd.sdq.probfunction.math.IProbabilityFunctionFactory;
 import de.uka.ipd.sdq.probfunction.math.IRandomGenerator;
 import de.uka.ipd.sdq.probfunction.math.impl.ProbabilityFunctionFactoryImpl;
-import de.uka.ipd.sdq.simucomframework.SimuComDefaultRandomNumberGenerator;
+import de.uka.ipd.sdq.simucomframework.core.SimuComDefaultRandomNumberGenerator;
 import de.uka.ipd.sdq.simucomframework.variables.StackContext;
 import de.uka.ipd.sdq.simucomframework.variables.cache.StoExCache;
 import de.uka.ipd.sdq.simucomframework.variables.cache.StoExCacheEntry;
